Crate picolo

Picolo is a simple libary for generating all the data prescribed in an image: The dimensions of the image and all pixels

Example

Basic usage for reading and translating a complete Image and translating it to a comprehensive table:

let pixl_struct = load_picture("images/ig-icon.png");

for i in pixl_struct {
    println!("{:?}", i); 
}

Structs

Color

Rgba colors

Pixel

Pixel struct that also derives Color struct

Functions

get_height

Returns height of the image as u32

get_width

Returns the width of image as u32

load_picture

Takes a borrowed string slice Str as its path.